+++ packages/asm/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "asm/5.0.0-beta.157";
+export declare const version = "asm/5.0.0-beta.158";