+++ packages/contracts/lib.esm/index.d.ts
@@ -25,6 +25,7 @@ export interface Event extends Log {
     event?: string;
     eventSignature?: string;
     args?: Result;
+    decodeError?: Error;
     decode?: (data: string, topics?: Array<string>) => any;
     removeListener: () => void;
     getBlock: () => Promise<Block>;
@@ -52,6 +53,7 @@ declare class RunningEvent {
     listenerCount(): number;
     run(args: Array<any>): number;
     prepareEvent(event: Event): void;
+    getEmit(event: Event): Array<any>;
 }
 export declare type ContractInterface = string | Array<Fragment | JsonFragment | string> | Interface;
 export declare class Contract {
@@ -85,7 +87,7 @@ export declare class Contract {
     private _normalizeRunningEvent;
     private _getRunningEvent;
     _checkRunningEvents(runningEvent: RunningEvent): void;
-    private _wrapEvent;
+    _wrapEvent(runningEvent: RunningEvent, log: Log, listener: Listener): Event;
     private _addEventListener;
     queryFilter(event: EventFilter, fromBlockOrBlockhash?: BlockTag | string, toBlock?: BlockTag): Promise<Array<Event>>;
     on(event: EventFilter | string, listener: Listener): this;