+++ packages/contracts/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "contracts/5.0.0-beta.147";
+export declare const version = "contracts/5.0.0-beta.148";